June 2024-2 Updates

See merge request dps8m/dps8m!409

+++ src/simh/scp.c
@@ -4559,7 +4559,7 @@ static void strip_spaces(char* str) {
   int i, x;
   for (i=x=0; str[i]; ++i)
     {
-      if (!isspace(str[i]) || (i > 0 && !isspace(str[(int)(i-1)]))) //-V781
+      if (!isspace((int)str[i]) || (i > 0 && !isspace((int)str[(int)(i-1)]))) //-V781
         {
           str[x++] = str[i];
         }
